pyproject.toml
[tool.poetry]
name = "azdfs"
version = "0.1.0"
description = ""
authors = ["Alex Zvorygin <grafetu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pysyncobj = "^0.3.10"
toml = "^0.10.2"
refuse = "^0.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
